fix(ui): 显式转换查找高亮矩形为 qreal，修复 macOS CI 的 -Wdouble-promotion 失败
QRectF 的实参类型是 qreal（macOS/64 位平台上即 double），而
renderer::FindMatch 的字段为 float，隐式 float→double 提升在
NEKO_WARNINGS_AS_ERRORS=ON 的 CI 中是错误。

行为等价，仅消除隐式提升；已在本地用 clang + 项目警告集验证。

+++ src/ui/src/web_view.cpp
@@ -533,10 +533,13 @@ void WebView::PaintFindHighlight(QPainter& painter)
   if (match.width <= 0.0F || match.height <= 0.0F) {
     return;
   }
-  const QRectF rect(match.x,
-                    match.y - static_cast<float>(verticalScrollBar()->value()),
-                    match.width,
-                    match.height);
+  // FindMatch holds float (device px); QRectF is qreal (double on some
+  // platforms).  Convert explicitly: an implicit float -> double conversion
+  // would trip -Wdouble-promotion, which is an error in CI.
+  const QRectF rect(static_cast<qreal>(match.x),
+                    static_cast<qreal>(match.y) - static_cast<qreal>(verticalScrollBar()->value()),
+                    static_cast<qreal>(match.width),
+                    static_cast<qreal>(match.height));
   painter.fillRect(rect, QColor(255, 214, 0, 128));
   painter.setPen(QColor(180, 140, 0, 200));
   painter.drawRect(rect.adjusted(0, 0, -1, -1));
